message batching into dev

+++ src/bot/commands/clearchannel.js
@@ -3,6 +3,9 @@ const clearEventByID = require('../../db/interfaces/postgres/update').clearEvent
 
 module.exports = {
   func: async message => {
+    if (!message.channel.guild.members.get(global.bot.user.id).permission.json['sendMessages']) {
+      return
+    }
     await webhookCache.deleteWebhook(message.channel.id)
     await clearEventByID(message.channel.guild.id, message.channel.id) // any event logging to this channel id will be wiped
     try {
